lambda/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.34
PyJWT==2.8.0
cryptography==42.0.0
requests==2.31.0
